Decompose game details
